IGNITE-2616 NonHeap memory usage metrics don't work as expected.
